+++ src/LinkAppenderLanguages/IN.ts
@@ -3,10 +3,9 @@ import { ILinkAppender } from '../ILinkAppender';
 export class LinkAppenderIN implements ILinkAppender {
 
     public label: string = "Viewing History";
-    public viewingHistoryUrl: string = "https://www.amazon.co.jp/gp/video/settings/watch-history/ref=atv_set_watch-history";
+    public viewingHistoryUrl: string = "https://www.primevideo.com/region/fe/settings/watch-history/ref=atv_set_watch-history";
 
     public addForLoggedInPage() {
-
         const parentElement = document.querySelector('#nav-xshop');
 
         if (!parentElement) {